Product Recommendation and Sentiment analysis on Amazon Review Dataset available at http://snap.stanford.edu/data/web-Amazon.html
In our project we are taking into consideration the amazon review dataset for Clothes, shoes and jewelleries and Beauty products. We are considering the reviews and ratings given by the user to different products as well as his/her reviews about his/her experience with the product(s).

Based on these input factors, sentiment analysis is performed on predicting the helpfulness of the reviews. Moreover, we also designed item-based collaborative filtering model based on k-Nearest Neighbors to find the 2 most similar items.